Add component and style counters.
