+++ assets/less/main.less
@@ -258,6 +258,26 @@ h6 {
   }
 }
 
+.block-counter {
+  background-color:#f5f5f5;
+  padding: 20px 0;
+  border-radius: 5px;
+  .count-title {
+      font-size: 40px;
+      font-weight: normal;
+      margin-top: 10px;
+      margin-bottom: 0;
+      text-align: center;
+  }
+  .count-text {
+      font-size: 13px;
+      font-weight: normal;
+      margin-top: 10px;
+      margin-bottom: 0;
+      text-align: center;
+  }
+}
+
 @media(max-width:991px) {
   #home-claim {
     background-position: center;
@@ -657,8 +677,8 @@ p.copyright {
 #home-days {
   .gradient(to left, @primary_color,  @secondary_color);
   color: @lightgrey_color;
-  .daysnumber {
-    font-size: 7em;
+  .daysnumber, h2 {
+    font-size: 50px;
     font-weight: bold;
     .title-letter-spacing;
     padding-right: 40px;
